feats/improving-test

+++ src/main/java/br/com/student/portal/service/TaskService.java
@@ -1,6 +1,7 @@
 package br.com.student.portal.service;
 
 import br.com.student.portal.entity.TaskEntity;
+import br.com.student.portal.exception.ObjectNotFoundException;
 import br.com.student.portal.repository.TaskRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -17,7 +18,9 @@ public class TaskService {
     private final TaskRepository taskRepository;
 
     public Optional<TaskEntity> findTaskById(UUID id) {
-        //TODO: Ao buscar uma tarefa inexistente devemos retornar uma mensagem de erro
+        if (taskRepository.findById(id).isEmpty()) {
+            throw new ObjectNotFoundException("There's no task with this id");
+        }
         return taskRepository.findById(id);
 
     }
@@ -30,7 +33,9 @@ public class TaskService {
     }
 
     public void taskDelete(UUID id) {
-        //TODO: Ao deletar uma tarefa inexistente devemos retornar uma mensagem de erro
+        if (taskRepository.findById(id).isEmpty()) {
+            throw new ObjectNotFoundException("There's no task with this id");
+        }
         taskRepository.deleteById(id);
     }
